Move Binomial distribution upstream
