+++ pyro/distributions/__init__.py
@@ -1,7 +1,6 @@
 from __future__ import absolute_import, division, print_function
 
 import pyro.distributions.torch_patch  # noqa F403
-from pyro.distributions.binomial import Binomial
 from pyro.distributions.delta import Delta
 from pyro.distributions.distribution import Distribution
 from pyro.distributions.empirical import Empirical
@@ -26,7 +25,6 @@ __all__ = [
     "is_validation_enabled",
     "validation_enabled",
     "AVFMultivariateNormal",
-    "Binomial",
     "Delta",
     "Distribution",
     "Empirical",
